Fix to replace None by "someone" when replicating
When replicating, if the order does not contain a prefferred name, the
tool cashed (trying to use an array accessor on None).
With this fix we check if prefferred name is None, and if so we use the
string "someone" (anonymous). If the array contains an answer, the given
preferred name will be used instead.