adding a simple test for the lottery registration
obs: it still doesn't work properly, under the setup, but that is
because we need some fixture data for a dummy borderland event.

Note also that these tests are really heavy and slow to load and start.
It is not worthy to have just one test, but one step at a time.